+++ CHANGELOG.md
@@ -14,6 +14,7 @@ this does, see the README.md entry.
 - Infrastructure required to do community installs of ansible modules to include SOPS
 - Added support for DEBUG_WAIT_TIME
 - Added support for special debug playbooks
+- Added session label for ara recording
 
 ### Changed
 - Disabled concurrent Jenkins builds on same branch/commit